refactor: refactoring controller and DTO code more validations and response

+++ src/main/java/com/daniel/menu/controller/FoodController.java
@@ -6,6 +6,7 @@ import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
@@ -14,6 +15,7 @@ import com.daniel.menu.dto.*;
 import com.daniel.menu.entity.Food;
 import com.daniel.menu.repository.FoodRepository;
 
+import jakarta.validation.Valid;
 import lombok.var;
 
 
@@ -34,32 +36,39 @@ public class FoodController {
     }
 
     @PostMapping
-    public void saveNewFood(@RequestBody FoodRequestDTO data) {
+    public ResponseEntity<Food> saveNewFood(@Valid @RequestBody FoodRequestDTO data) {
         Food foodData = new Food(data);
         repository.save(foodData);
+        return ResponseEntity.status(HttpStatus.CREATED).body(foodData);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteFood(@PathVariable Long id) {
+    public ResponseEntity<?> deleteFood(@PathVariable Long id) {
+        if (!repository.existsById(id)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Comida não encontrada");
+        }
         repository.deleteById(id);
+        return ResponseEntity.noContent().build();
     }
 
-    @PutMapping("/{name}/{id}")
-    public void putFood(@PathVariable String name, @PathVariable Long id, @RequestBody FoodRequestDTO data) {
+    @PutMapping("/{id}")
+    public ResponseEntity<?> updateFood(@PathVariable Long id, @Valid @RequestBody FoodRequestDTO data) {
         var specificFood = repository.findById(id);
 
-        if (specificFood.isPresent()) {
-            Food existingFood = specificFood.get();
-    
-            existingFood.setTitle(data.title());
-            existingFood.setImage(data.image());
-            existingFood.setPrice(data.price());
-    
-            repository.save(existingFood);
-    
-        } 
+        if (specificFood.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Comida não encontrada!");
+        }
+
+        Food existingFood = specificFood.get();
+        existingFood.setTitle(data.title());
+        existingFood.setImage(data.image());
+        existingFood.setPrice(data.price());
+
+        repository.save(existingFood);
         
+        return ResponseEntity.ok(existingFood);
     }
+
 
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     @ExceptionHandler(MethodArgumentNotValidException.class)
